+++ R/create.R
@@ -672,13 +672,13 @@ createGiottoObjectSubcellular <- function(
       remove_background_polygon = TRUE,
       background_algo = c("range"),
       fill_holes = TRUE,
+      ID_fmt = "cell_",
       poly_IDs = NULL,
       flip_vertical = TRUE,
       shift_vertical_step = TRUE,
       flip_horizontal = TRUE,
       shift_horizontal_step = TRUE,
-      calc_centroids = FALSE,
-      fix_multipart = TRUE
+      calc_centroids = FALSE
     )
   }
 
@@ -1917,7 +1917,7 @@ create_giotto_points_object <- function(feat_type = "rna",
 #' spatial annotations and polygons. Inputs can be from a structured data.frame
 #' object where three of the columns should correspond to x/y vertices and the
 #' polygon ID and additional columns are set as attributes, a spatial file
-#' such as wkt, .shp, or .GeoJSON, or a mask file (e.g. segmentation results)
+#' such as wkt, .shp, or .GeoJSON, or a mask file (e.g. segmentation results).
 #' @param x input. Filepath to a .GeoJSON or a mask image file. Can also be a
 #' data.frame with vertex 'x', 'y', and 'poly_ID' information.
 #' @param name name for polygons
@@ -1956,17 +1956,6 @@ setMethod(
 )
 
 #' @rdname createGiottoPolygon
-#' @param mask_method how the mask file defines individual segmentation annotations
-#' @param remove_background_polygon try to remove background polygon (default: FALSE)
-#' @param background_algo algorithm to remove background polygon
-#' @param fill_holes fill holes within created polygons
-#' @param poly_IDs unique names for each polygon in the mask file
-#' @param flip_vertical flip mask figure in a vertical manner
-#' @param shift_vertical_step shift vertical (boolean or numerical)
-#' @param flip_horizontal flip mask figure in a horizontal manner
-#' @param shift_horizontal_step shift horizontal (boolean or numerical)
-#' @param fix_multipart try to split polygons with multiple parts (default: TRUE)
-#' @param remove_unvalid_polygons remove unvalid polygons (default: TRUE)
 #' @export
 setMethod(
   "createGiottoPolygon", signature("SpatRaster"),
@@ -1977,11 +1966,11 @@ setMethod(
            background_algo = c("range"),
            fill_holes = TRUE,
            poly_IDs = NULL,
+           ID_fmt = "cell_",
            flip_vertical = TRUE,
            shift_vertical_step = TRUE,
            flip_horizontal = TRUE,
            shift_horizontal_step = TRUE,
-           fix_multipart = TRUE,
            remove_unvalid_polygons = TRUE,
            calc_centroids = FALSE,
            verbose = TRUE) {
@@ -1995,18 +1984,85 @@ setMethod(
       background_algo = background_algo,
       fill_holes = fill_holes,
       poly_IDs = poly_IDs,
+      ID_fmt = ID_fmt,
       flip_vertical = flip_vertical,
       shift_vertical_step = shift_vertical_step,
       flip_horizontal = flip_horizontal,
       shift_horizontal_step = shift_horizontal_step,
-      fix_multipart = fix_multipart,
       remove_unvalid_polygons = remove_unvalid_polygons,
       calc_centroids = calc_centroids
     )
   }
 )
 
+
 #' @rdname createGiottoPolygon
+#' @param \dots additional params to pass. For character method, params pass to
+#' SpatRaster or SpatVector methods, depending on whether x was a filepath to
+#' a maskfile or a spatial file (ex: wkt, shp, GeoJSON) respectively.
+#' @examples
+#' # %%%%%%%%% `createGiottoPolygon()` examples %%%%%%%%% #
+#' # ------- create from a mask image ------- #
+#' m <- system.file("extdata/toy_mask_multi.tif", package = "GiottoClass")
+#' plot(terra::rast(m), col = grDevices::hcl.colors(7))
+#' gp <- createGiottoPolygon(
+#'   m,
+#'   flip_vertical = FALSE, flip_horizontal = FALSE,
+#'   shift_horizontal_step = FALSE, shift_vertical_step = FALSE,
+#'   ID_fmt = "id_test_%03d",
+#'   name = "test"
+#' )
+#' plot(gp, col = grDevices::hcl.colors(7))
+#'
+#' # ------- create from an shp file ------- #
+#' shp <- system.file("extdata/toy_poly.shp", package = "GiottoClass")
+#' # vector inputs do not have params for flipping and shifting
+#' gp2 <- createGiottoPolygon(shp, name = "test")
+#' plot(gp2, col = grDevices::hcl.colors(7))
+#'
+#'
+#' @export
+setMethod(
+  "createGiottoPolygon", signature("character"),
+  function(x, ...) {
+    checkmate::assert_file_exists(x)
+
+    # try success means it should be mask file
+    # try failure means it should be vector file
+    try_rast <- tryCatch(
+      {
+        terra::rast(x)
+      },
+      error = function(e) return(invisible(NULL)),
+      warning = function(w) {NULL}
+    )
+
+    # mask workflow
+    if (inherits(try_rast, "SpatRaster")) {
+      return(createGiottoPolygon(try_rast, ...))
+    }
+
+    # file workflow
+    return(createGiottoPolygon(x = terra::vect(x), ...))
+  }
+)
+
+
+#' @rdname createGiottoPolygon
+#' @examples
+#' # ------- create from data.frame-like ------- #
+#' shp <- system.file("extdata/toy_poly.shp", package = "GiottoClass")
+#' gpoly <- createGiottoPolygon(shp, name = "test")
+#' plot(gpoly)
+#' gpoly_dt <- data.table::as.data.table(gpoly, geom = "XY")
+#' needed_cols_dt <- gpoly_dt[, .(geom, part, x, y, hole, poly_ID)]
+#' force(needed_cols_dt)
+#'
+#' out <- createGiottoPolygon(needed_cols_dt,
+#'                            name = "test")
+#' plot(out)
+#'
+#'
 #' @export
 setMethod(
   "createGiottoPolygon", signature("data.frame"),
@@ -2027,74 +2083,97 @@ setMethod(
   }
 )
 
-#' @rdname createGiottoPolygon
-#' @param \dots additional params to pass. For character method, params pass to
-#' SpatRaster or SpatVector methods, depending on whether x was a filepath to
-#' a maskfile or a spatial file (ex: wkt, shp, GeoJSON) respectively.
-#' @export
-setMethod(
-  "createGiottoPolygon", signature("character"),
-  function(x, ...) {
-    checkmate::assert_file_exists(x)
 
-    # try success means it should be mask file
-    # try failure means it should be vector file
-    try_rast <- try(
-      {
-        terra::rast(x)
-      },
-      silent = TRUE
-    )
-
-    # mask workflow
-    if (inherits(try_rast, "SpatRaster")) {
-      return(createGiottoPolygon(x, ...))
-    }
-
-    # file workflow
-    return(createGiottoPolygon(
-      x = terra::vect(x),
-      ...
-    ))
-  }
-)
-
-
-#' @title Create giotto polygons from mask file
 #' @rdname createGiottoPolygon
 #' @param maskfile path to mask file
-#' @param mask_method how the mask file defines individual segmentation annotations
-#' @param name name for polygons
+#' @param mask_method how the mask file defines individual segmentation annotations.
+#' See *mask_method* section
+#' @param name character. Name to assign created `giottoPolygon`
 #' @param remove_background_polygon try to remove background polygon (default: FALSE)
 #' @param background_algo algorithm to remove background polygon
 #' @param fill_holes fill holes within created polygons
-#' @param poly_IDs unique names for each polygon in the mask file
+#' @param poly_IDs character vector. Default = NULL. Custom unique names for
+#' each polygon in the mask file.
+#' @param ID_fmt character. Only applied if `poly_IDs = NULL`. Naming scheme for
+#' poly_IDs. Default = "cell_". See *ID_fmt* section.
 #' @param flip_vertical flip mask figure in a vertical manner
 #' @param shift_vertical_step shift vertical (boolean or numerical)
 #' @param flip_horizontal flip mask figure in a horizontal manner
 #' @param shift_horizontal_step shift horizontal (boolean or numerical)
-#' @param calc_centroids calculate centroids for polygons
-#' @param fix_multipart try to split polygons with multiple parts (default: TRUE)
 #' @param remove_unvalid_polygons remove unvalid polygons (default: TRUE)
-#' @return a giotto polygon object
 #' @concept mask polygon
+#' @section mask_method:
+#' One of "single", "multiple", or "guess".
+#' \itemize{
+#'   \item{*"single"* assumes that the provided mask image is binary, with only
+#'   polygon vs background being distinct values. With this kind of image, the
+#'   expected generated polygons is a single multipart polygon. "single" takes
+#'   this multipart polygon and breaks it apart into individual singlepart
+#'   polygons. An initial simple `numeric` index as the 'nth' polygon found in
+#'   the mask image will be applied as an ID (see *ID_fmt* section).}
+#'   \item{*"multiple"* assumes that the provided mask image has distinct
+#'   intensity values to specify the IDs of individual polygons. An initial
+#'   `numeric` ID is applied as the intensity value of the pixels that made up
+#'   the annotation for that polygon in the mask image (see *ID_fmt* section).}
+#'   \item{*"guess"* examines the values in the image to pick the most likely
+#'   appropriate method out of "single" or "multiple".}
+#' }
+#' @section ID_fmt:
+#' Defaults to applying the input as a prefix (using `paste0()`) to the
+#' numerical ID values detected by  `mask_method`. (ie: `ID_fmt = "cell_"`
+#' produces `cell_1`, `cell_2`, `cell_3`, ...)\cr
+#' If a "%" character is detected in the input then the input will be treated as
+#' a `sprintf()` `fmt` param input instead. (ie: `ID_fmt = "cell_%03d"` produces
+#' `cell_001`, `cell_002`, `cell_003`, ...)
+#' @return a giotto polygon object
+#' @examples
+#' # %%%%%%%%% `createGiottoPolygonsFromMask()` examples %%%%%%%%% #
+#' mask_multi <- system.file("extdata/toy_mask_multi.tif",
+#'                           package = "GiottoClass")
+#' mask_single <- system.file("extdata/toy_mask_single.tif",
+#'                            package = "GiottoClass")
+#' plot(terra::rast(mask_multi), col = grDevices::hcl.colors(7))
+#' plot(terra::rast(mask_single))
+#'
+#' gpoly1 = createGiottoPolygonsFromMask(
+#'   mask_multi,
+#'   flip_vertical = FALSE, flip_horizontal = FALSE,
+#'   shift_horizontal_step = FALSE, shift_vertical_step = FALSE,
+#'   ID_fmt = "id_test_%03d",
+#'   name = "multi_test"
+#' )
+#' plot(gpoly1, col = grDevices::hcl.colors(7))
+#'
+#' gpoly2 = createGiottoPolygonsFromMask(
+#'   mask_single,
+#'   flip_vertical = FALSE, flip_horizontal = FALSE,
+#'   shift_horizontal_step = FALSE, shift_vertical_step = FALSE,
+#'   ID_fmt = "id_test_%03d",
+#'   name = "single_test"
+#' )
+#' plot(gpoly2, col = grDevices::hcl.colors(5))
 #' @export
-createGiottoPolygonsFromMask <- function(maskfile,
-                                         mask_method = c("guess", "single", "multiple"),
-                                         name = "cell",
-                                         remove_background_polygon = FALSE,
-                                         background_algo = c("range"),
-                                         fill_holes = TRUE,
-                                         poly_IDs = NULL,
-                                         flip_vertical = TRUE,
-                                         shift_vertical_step = TRUE,
-                                         flip_horizontal = TRUE,
-                                         shift_horizontal_step = TRUE,
-                                         calc_centroids = FALSE,
-                                         fix_multipart = TRUE,
-                                         remove_unvalid_polygons = TRUE) {
+createGiottoPolygonsFromMask <- function(
+    maskfile,
+    mask_method = c("guess", "single", "multiple"),
+    name = "cell",
+    remove_background_polygon = FALSE,
+    background_algo = c("range"),
+    fill_holes = TRUE,
+    poly_IDs = NULL,
+    ID_fmt = "cell_",
+    flip_vertical = TRUE,
+    shift_vertical_step = TRUE,
+    flip_horizontal = TRUE,
+    shift_horizontal_step = TRUE,
+    calc_centroids = FALSE,
+    remove_unvalid_polygons = TRUE,
+    verbose = FALSE
+) {
   # data.table vars
   x <- y <- geom <- part <- NULL
+
+  remove_unvalid_polygons <- as.logical(remove_unvalid_polygons)
 
   # select background algo
   background_algo <- match.arg(background_algo, choices = "range")
@@ -2120,57 +2199,88 @@ createGiottoPolygonsFromMask <- function(maskfile,
 
   # create polygons from mask
   rast_dimensions <- dim(terra_rast)
+  # value = TRUE here means that the intensity value of the mask image
+  # (which usually encodes the intended polygon ID) is added to the resulting
+  # SpatVector as the only attribute.
   terra_polygon <- terra::as.polygons(x = terra_rast, value = TRUE)
+  val_col <- names(terra_polygon) # the only col should be from the values
 
   # fill holes
   if (isTRUE(fill_holes)) {
     terra_polygon <- terra::fillHoles(terra_polygon)
   }
 
-  # remove unvalid polygons
-  if (isTRUE(remove_unvalid_polygons)) {
+  # handle unvalid polygons ##
+  # The unvalid polys formed from as.polygons are usually very misshapen and
+  # artefacted. It is impossible to fix them using `terra::makeValid()`
+  if (remove_unvalid_polygons) {
     valid_index <- terra::is.valid(terra_polygon)
     terra_polygon <- terra_polygon[valid_index]
   }
 
-
-  spatVecDT <- .spatvector_to_dt(terra_polygon)
-
   ## flip across axes ##
   if (isTRUE(flip_vertical)) {
-    # terra_polygon = terra::flip(terra_polygon, direction = 'vertical')
-    spatVecDT[, y := -y]
+    terra_polygon <- .flip_spatvect(terra_polygon)
   }
-
   if (isTRUE(flip_horizontal)) {
-    # terra_polygon = terra::flip(terra_polygon, direction = 'horizontal')
-    spatVecDT[, x := -x]
+    terra_polygon <- .flip_spatvect(terra_polygon)
   }
 
-  # guess mask method
+  # convert to DT format since we want to be able to compare number of geoms
+  # vs polys to determine correct mask method.
+  # TODO only test a subset of polys here?
+  spatVecDT <- .spatvector_to_dt(terra_polygon)
+
+  ## guess mask method ##
   if (mask_method == "guess") {
     uniq_geoms <- length(unique(spatVecDT$geom))
     uniq_parts <- length(unique(spatVecDT$part))
     mask_method <- ifelse(uniq_geoms > uniq_parts, "multiple", "single")
   }
+  vmsg(.v = verbose, sprintf("parsing mask using mask_method: %s", mask_method))
 
-  if (mask_method == "multiple") {
-    if (is.null(poly_IDs)) {
-      spatVecDT[, geom := paste0(name, geom)]
+
+  ## define polys and apply auto IDs ##
+  naming_fun <- ifelse(grepl("%", ID_fmt), sprintf, paste0)
+  # If poly_IDs are NOT provided, then terra_polygon IDs created here will be
+  # `character` and the finalized ID values.
+  # If poly_IDs ARE provided, the IDs are still temporary and MUST remain
+  # `numeric`, pending the `poly_IDs` param being applied downstream.
+  terra_polygon <- switch(mask_method,
+    "multiple" = {
+      names(terra_polygon) <- "poly_ID"
+      if (is.null(poly_IDs)) {
+        # spatVecDT[, geom := naming_fun(ID_fmt, geom)]
+        # spatVecDT[, (val_col) := naming_fun(ID_fmt, get(val_col))]
+        # g_polygon <- createGiottoPolygonsFromDfr(
+        #   segmdfr = spatVecDT[, .(x, y, get(val_col))]
+        # )
+        # g_polygon@spatVector
+        terra_polygon$poly_ID <- naming_fun(ID_fmt, terra_polygon$poly_ID)
+      }
+      terra_polygon
+    },
+    "single" = {
+      # TODO ordering may be performed based on centroids xy instead of
+      # converting the full polygon and then ordering on parts
+      # May improve the speed
+      if (is.null(poly_IDs)) {
+        spatVecDT[, part := naming_fun(ID_fmt, part)]
+      }
+      g_polygon <- createGiottoPolygonsFromDfr(
+        segmdfr = spatVecDT[, .(x, y, part)]
+      )
+      if (!is.null(poly_IDs)) {
+        g_polygon@spatVector$poly_ID <- as.numeric(g_polygon@spatVector$poly_ID)
+      }
+
+      g_polygon@spatVector
     }
-    g_polygon <- createGiottoPolygonsFromDfr(segmdfr = spatVecDT[, .(x, y, geom)])
-    terra_polygon <- g_polygon@spatVector
-  } else if (mask_method == "single") {
-    if (is.null(poly_IDs)) {
-      spatVecDT[, part := paste0(name, part)]
-    }
-    g_polygon <- createGiottoPolygonsFromDfr(segmdfr = spatVecDT[, .(x, y, part)])
-    terra_polygon <- g_polygon@spatVector
-  }
+  )
 
 
 
-  ## shift values ##
+  ## apply spatial shifts ##
   if (identical(shift_vertical_step, TRUE)) {
     shift_vertical_step <- rast_dimensions[1] # nrows of raster
   } else if (is.numeric(shift_vertical_step)) {
@@ -2193,29 +2303,35 @@ createGiottoPolygonsFromMask <- function(maskfile,
   )
 
 
-  # remove background polygon
+  ## remove background polygon ##
   if (isTRUE(remove_background_polygon)) {
     if (background_algo == "range") {
       backgr_poly_id <- .identify_background_range_polygons(terra_polygon)
-      # print(backgr_poly_id) uneccessary to print?
+      vmsg(.v = verbose, sprintf("removed background poly.\n ID was: %s",
+                                 backgr_poly_id))
     }
 
-    terra_polygon <- terra::subset(x = terra_polygon, terra_polygon[["poly_ID"]] != backgr_poly_id)
+    terra_polygon <- terra::subset(
+      x = terra_polygon,
+      terra_polygon[["poly_ID"]] != backgr_poly_id
+    )
   }
 
 
-  # provide own cell_ID name
+  ## apply custom poly_IDs ##
   if (!is.null(poly_IDs)) {
+    # first sort the polys by ID to ensure that custom poly_IDs are applied in a
+    # meaningful manner
+    terra_polygon <- terra_polygon[order(terra_polygon$poly_ID)]
+
     if (isTRUE(remove_unvalid_polygons)) {
       poly_IDs <- poly_IDs[valid_index]
     }
 
     if (length(poly_IDs) != nrow(terra::values(terra_polygon))) {
-      stop("length cell_IDs does not equal number of found polyongs \n")
+      stop("length cell_IDs does not equal number of found polygons \n")
     }
     terra_polygon$poly_ID <- as.character(poly_IDs)
-  } else {
-    terra_polygon$poly_ID <- paste0(name, "_", 1:nrow(terra::values(terra_polygon)))
   }
 
 
@@ -2254,7 +2370,7 @@ createGiottoPolygonsFromMask <- function(maskfile,
 #' @param copy_dt (default TRUE) if segmdfr is provided as dt, this determines
 #' whether a copy is made
 #' @param verbose be verbose
-#' @details When determining which column within the tabular data is intended to
+#' @details When determining which column within tabular data is intended to
 #' provide polygon information, Giotto first checks the column names for 'x', 'y',
 #' and 'poly_ID'. If any of these are discovered, they are directly selected. If
 #' this is not discovered then Giotto checks the data type of the columns and selects